requirements.txt
concurrent.futures
oslo.config
pyzabbix
requests
yaml
